Fix quant specific op registration for some ops
Quant specific registration for BatchMatmul, SVDF and LSTM were not
working correctly.
